This doesn't need to be a trait anymore
